💥 Split Text component into Text and TextStyle
Migration Guide:

The `Text` component was previously used both for adding blocks of text inside stacks and styling just a span of text inside a larger text block. Using the same component for both these behaviors could lead to some strange behavior on some platforms. The `Text` component is now used only for blocks of text, and a new component `TextStyle` has been introduced for applying styles to an inline span of text.